ECS: Introduce Entity Component System
This rough first iteration suffers from the fact it is based on parallel lists made of Entities & Componenets. So to find a single Entity's components that it has being tracked, you must iterate through *every* entity in the component list to find out if it's in there. Instead, the future system will use an EntityIndex to directly index into the component list _safely_ allowing better accessing of ALL necessary components for entity-entity interactions.